% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_comparativeServer.R
\name{comparativeServer}
\alias{comparativeServer}
\title{comparativeServer}
\usage{
comparativeServer(input, output, session, metadata_rv, timeseries_rv)
}
\arguments{
\item{input}{module}

\item{output}{module}

\item{session}{module}

\item{metadata_rv}{reactive values returned from metadata tab}

\item{timeseries_rv}{reactive values returned from timeseries tab}
}
\description{
comparativeServer
}
