% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_timeseriesServer.R
\name{timeseriesServer}
\alias{timeseriesServer}
\title{timeseriesServer}
\usage{
timeseriesServer(input, output, session, x, metadata_rv)
}
\arguments{
\item{input}{module}

\item{output}{module}

\item{session}{module}

\item{x}{used to separately pass session to updateNavBarPage}

\item{metadata_rv}{reactive values returned from metadata tab}

\item{comparative_results}{results from comparative tab}
}
\description{
timeseriesServer
}
