% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_metadataServer.R
\name{metadataServer}
\alias{metadataServer}
\title{metadataServer}
\usage{
metadataServer(input, output, session, x)
}
\arguments{
\item{input}{module}

\item{output}{module}

\item{session}{module}

\item{x}{used to separately pass session to updateNavBarPage}
}
\description{
metadataServer
}
